# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: UltraV API
  version: 1.0.0
  summary: API-first KVM/QEMU virtual machine management
  description: |
    UltraV is a modern, simple, API-first management layer for KVM/QEMU
    virtual machines on single Linux hosts.

    **Everything is an API.** Every UI capability is available here first.
    This contract is the source of truth for the Web UI, CLI, scripts,
    automation, and future AI agents (MCP adapters).

    ## Conventions
    - All requests/responses are JSON.
    - Every response includes an `X-Request-Id` header; errors also include `requestId` in the body.
    - Errors use a stable envelope: `{"error": {"code", "message", "requestId"}}`.
    - Destructive or long-running operations will use specific, semantically
      named operations (never generic `execute`/`command` endpoints).
  license:
    name: TBD
servers:
  - url: /api/v1
    description: Local host (single-node)
...
paths:
  /auth/login:
    post:
      operationId: login
      tags: [Auth]
      summary: Authenticate and obtain a bearer token
      description: >
        Validates credentials against the local user database (SQLite) and
        returns a signed JWT bearer token. All other endpoints (except
        /health and /ready) require `Authorization: Bearer <token>`.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
            example: { username: admin, password: secret }
      responses:
        '200':
          description: Authenticated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LoginResponse' }
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              example: { error: { code: INVALID_CREDENTIALS, message: Invalid username or password, requestId: req_x } }
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [Auth]
      summary: Get the authenticated user
      responses:
        '200':
          description: Current user
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /auth/password:
    put:
      operationId: changeOwnPassword
      tags: [Auth]
      summary: Change the authenticated user's own password
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ChangePasswordRequest' }
      responses:
        '204': { description: Password changed }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /users:
    get:
      operationId: listUsers
      tags: [Users]
      summary: List local users
      description: Requires the `admin` role.
      responses:
        '200':
          description: User list
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserList' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      operationId: createUser
      tags: [Users]
      summary: Create a local user
      description: Requires the `admin` role.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateUserRequest' }
      responses:
        '201':
          description: Created user
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409':
          description: Username already exists
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              example: { error: { code: USER_ALREADY_EXISTS, message: A user with this username already exists, requestId: req_x } }
  /users/{id}:
    delete:
      operationId: deleteUser
      tags: [Users]
      summary: Delete a local user
      description: >
        Requires the `admin` role. The caller cannot delete itself and the
        last remaining admin cannot be deleted.
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '204': { description: User deleted }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404':
          description: User not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '409':
          description: Cannot delete self or the last admin
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              example: { error: { code: LAST_ADMIN, message: Cannot delete the last admin user, requestId: req_x } }
  /users/{id}/password:
    put:
      operationId: resetUserPassword
      tags: [Users]
      summary: Reset a user's password (admin)
      description: Requires the `admin` role.
      parameters:
        - $ref: '#/components/parameters/UserId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdatePasswordRequest' }
      responses:
        '204': { description: Password updated }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404':
          description: User not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /host:
    get:
      operationId: getHost
      tags: [Host]
      summary: Get host status
      description: Returns status and identity of the hypervisor host.
      responses:
        '200':
          description: Host status
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Host' }
              example:
                hostname: kvm01
                operatingSystem: Ubuntu 24.04 LTS
                kernel: 6.8.0-45-generic
                uptimeSeconds: 128934
                cpu: { model: AMD EPYC 7402P 24-Core Processor, cores: 24, threads: 48, usagePercent: 23.5 }
                memoryTotalBytes: 68719476736
                memoryUsedBytes: 21474836480
                virtualization: { kvmEnabled: true }
                qemuVersion: 8.2.0
                libvirtVersion: 10.0.0
        '500': { $ref: '#/components/responses/InternalError' }
  /capabilities:
    get:
      operationId: getCapabilities
      tags: [Host]
      summary: Get host capabilities
      description: >
        Feature discovery for UI, automation and agents. Clients must use
        this to know what the host supports instead of hardcoding assumptions.
      responses:
        '200':
          description: Host capabilities
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Capabilities' }
              example:
                virtualization: { kvm: true, uefi: true }
                storage: { qcow2: true, raw: true, snapshots: true }
                features: { backup: false, migration: false, cloudInit: false }
        '500': { $ref: '#/components/responses/InternalError' }
...
  /storage/pools:
    get:
      operationId: listStoragePools
      tags: [Storage]
      summary: List storage pools
      responses:
        '200':
          description: Storage pools
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StoragePoolList' }
        '500': { $ref: '#/components/responses/InternalError' }
    post:
      operationId: createStoragePool
      tags: [Storage]
      summary: Create a storage pool
      description: |
        Defines, builds and starts a new directory-based storage pool.
        The target directory is created if it does not exist.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/StoragePoolCreate' }
            example: { name: iso-store, targetPath: /var/lib/libvirt/isos, autostart: true }
      responses:
        '201':
          description: Pool created (returns the pool)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StoragePool' }
        '400':
          description: Validation error (name/targetPath)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '409':
          description: A pool with this name already exists
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              example: { error: { code: STORAGE_POOL_ALREADY_EXISTS, message: A storage pool with this name already exists, requestId: req_01HV3M9Z2K2Q } }
        '500': { $ref: '#/components/responses/InternalError' }
  /storage/pools/{id}:
    get:
      operationId: getStoragePool
      tags: [Storage]
      summary: Get a storage pool
      parameters: [ { $ref: '#/components/parameters/ResourceID' } ]
      responses:
        '200':
          description: Storage pool details
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StoragePool' }
        '404':
          description: Storage pool not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '500': { $ref: '#/components/responses/InternalError' }
    delete:
      operationId: deleteStoragePool
      tags: [Storage]
      summary: Remove a storage pool from the listing
      description: >
        Undefines the storage pool (deactivating it first if needed) so it disappears from
        the listing. No data is deleted: volumes, disk images and files on disk are kept intact.
      parameters: [ { $ref: '#/components/parameters/ResourceID' } ]
      responses:
        '204': { description: Pool removed from the listing (data untouched) }
        '404':
          description: Storage pool not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '500': { $ref: '#/components/responses/InternalError' }
  /storage/pools/{id}/refresh:
    post:
      operationId: refreshStoragePool
      tags: [Storage]
      summary: Refresh a storage pool
      description: Re-syncs the pool's volume list and usage numbers with the backing storage.
      parameters: [ { $ref: '#/components/parameters/ResourceID' } ]
      responses:
        '200':
          description: Pool refreshed (returns updated pool)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StoragePool' }
        '404':
          description: Storage pool not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '500': { $ref: '#/components/responses/InternalError' }
...
security:
  - bearerAuth: []
components:
...
  responses:
...
    Unauthorized:
      description: Missing or invalid bearer token
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example:
            error: { code: UNAUTHORIZED, message: Authentication required, requestId: req_01HV3M9Z2N }
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username: { type: string, example: admin }
        password: { type: string, format: password, example: secret }
    LoginResponse:
      type: object
      required: [token, tokenType, expiresInSeconds, user]
      properties:
        token: { type: string, description: JWT bearer token, example: eyJhbGciOi... }
        tokenType: { type: string, example: Bearer }
        expiresInSeconds: { type: integer, example: 86400 }
        user: { $ref: '#/components/schemas/User' }
...